Face-Recognition-UsingPython.py
Machine Learning using Python

Face recognition is a biometric identification technique that uses unique characteristics of an individual's face to identify them. Most facial recognition systems work by comparing the face print to a database of known faces. If there's a match, the system can identify the individual. However, if the face print isn't in the database, the system can't identify the individual and it will return as Unknown. This project is done using “Open Source Computer Vision Library”, the OpenCV. OpenCV is the most popular library for computer vision. OpenCV uses machine learning algorithms to search for faces within a picture. Because faces are so complicated, there isn’t one simple test that will tell you if it found a face or not. Instead, there are thousands of small patterns and features that must be matched. The algorithms break the task of identifying the face into thousands of smaller, bite-sized tasks, each of which is easy to solve. For face detection, one of the most famous algorithms is known as the Viola-Jones Face Detection technique, commonly known as Haar cascades. A Haar cascade classifier is used to detect the image for which it has been trained.